+++ src/main/java/org/sosly/villagetale/capability/Capabilities.java
@@ -1,7 +1,6 @@
 package org.sosly.villagetale.capability;
 
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.chunk.LevelChunk;
 import net.minecraftforge.common.capabilities.Capability;
@@ -11,27 +10,19 @@ import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import org.sosly.villagetale.VillageTale;
-import org.sosly.villagetale.api.capability.IRecipeKnowledgeCapability;
 import org.sosly.villagetale.api.capability.IVillageCapability;
 import org.sosly.villagetale.api.capability.IVillagesCapability;
-import org.sosly.villagetale.capability.recipeknowledge.RecipeKnowledgeProvider;
 import org.sosly.villagetale.capability.village.VillageProvider;
 import org.sosly.villagetale.capability.villages.VillagesProvider;
-import org.sosly.villagetale.entity.Villager;
 
 @Mod.EventBusSubscriber(modid = VillageTale.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class Capabilities {
-    public static final Capability<IRecipeKnowledgeCapability> RECIPE_KNOWLEDGE_CAPABILITY =
-        CapabilityManager.get(new CapabilityToken<>() {});
-
     public static final Capability<IVillageCapability> VILLAGE_CAPABILITY =
         CapabilityManager.get(new CapabilityToken<>() {});
 
     public static final Capability<IVillagesCapability> VILLAGES_CAPABILITY =
         CapabilityManager.get(new CapabilityToken<>() {});
 
-    private static final ResourceLocation RECIPE_KNOWLEDGE_KEY =
-        new ResourceLocation(VillageTale.MOD_ID, "recipe_knowledge");
     private static final ResourceLocation VILLAGE_CAPABILITY_KEY =
         new ResourceLocation(VillageTale.MOD_ID, "village_capability");
 
@@ -47,17 +38,6 @@ public class Capabilities {
         village.setChunk(chunk);
         event.addCapability(VILLAGE_CAPABILITY_KEY, provider);
     }
-
-    @SubscribeEvent
-    public static void onAttachEntityCapabilities(AttachCapabilitiesEvent<Entity> event) {
-        if (!(event.getObject() instanceof Villager)) {
-            return;
-        }
-
-        RecipeKnowledgeProvider provider = new RecipeKnowledgeProvider();
-        event.addCapability(RECIPE_KNOWLEDGE_KEY, provider);
-    }
-
 
     @SubscribeEvent
     public static void onAttachLevelCapabilities(AttachCapabilitiesEvent<Level> event) {